DRTII-831 Move red list structures classes to shared
