+++ client/src/main/scala/drt/client/components/NationalityFinderComponent.scala
@@ -1,14 +1,13 @@
 package drt.client.components
 
-import drt.client.components.FlightTableRow.ApiIndirectRedListPax
 import drt.client.logger.{Logger, LoggerFactory}
-import drt.shared.{Nationality, RedList}
+import drt.shared.{ApiIndirectRedListPax, RedList}
 import japgolly.scalajs.react.component.Scala.Component
 import japgolly.scalajs.react.vdom.html_<^._
 import japgolly.scalajs.react.{CtorType, ScalaComponent}
 
 object NationalityFinderComponent {
-  def isRedListCountry(country: String): Boolean = redList.keys.exists(_.toLowerCase == country.toLowerCase)
+  def isRedListCountry(country: String): Boolean = RedList.countryToCode.keys.exists(_.toLowerCase == country.toLowerCase)
 
   val log: Logger = LoggerFactory.getLogger(getClass.getName)
   val component: Component[ApiIndirectRedListPax, Unit, Unit, CtorType.Props] = ScalaComponent.builder[ApiIndirectRedListPax]("FlightChart")
@@ -27,10 +26,6 @@ object NationalityFinderComponent {
     .build
 
   def apply(apiIndirectRedListPax: ApiIndirectRedListPax): VdomElement = component(apiIndirectRedListPax)
-
-  val redList: Map[String, String] = RedList.countryToCode
-
-  val redListNats: Iterable[Nationality] = redList.values.map(Nationality(_))
 
 }
